app/requirements.txt
cryptography>=41.0.7
fastapi>=0.86.0
gunicorn>=20.1.0
passlib>=1.7.4
passlib[bcrypt]
pycrypto>=2.6.1
pydantic>=1.10.2
pymysql>=1.1.0
python-dotenv>=1.0.0
python-jose[criptography]
uvicorn>=0.19.0